Move ValidationUtil to message_layer
