+++ src/main/java/com/streamr/client/protocol/message_layer/GroupKeyErrorResponse.java
@@ -1,6 +1,5 @@
 package com.streamr.client.protocol.message_layer;
 
-import com.streamr.client.utils.ValidationUtil;
 import java.util.List;
 import java.util.Objects;
