+++ src/main/java/com/streamr/client/protocol/message_layer/ValidationUtil.java
@@ -1,0 +1,17 @@
+package com.streamr.client.protocol.message_layer;
+
+import java.util.Collection;
+
+public class ValidationUtil {
+  public static void checkNotNull(Object value, String fieldName) {
+    if (value == null) {
+      throw new MalformedMessageException(fieldName + " can not be null");
+    }
+  }
+
+  public static <T> void checkNotEmpty(Collection<T> collection, String fieldName) {
+    if (collection.isEmpty()) {
+      throw new MalformedMessageException(fieldName + " can not be empty");
+    }
+  }
+}
