+++ src/main/java/com/streamr/client/utils/MessageCreationUtil.java
@@ -3,13 +3,13 @@ package com.streamr.client.utils;
 import com.streamr.client.exceptions.InvalidGroupKeyException;
 import com.streamr.client.exceptions.InvalidGroupKeyRequestException;
 import com.streamr.client.exceptions.InvalidGroupKeyResponseException;
-import com.streamr.client.protocol.message_layer.MalformedMessageException;
 import com.streamr.client.exceptions.SigningRequiredException;
 import com.streamr.client.protocol.message_layer.GroupKeyAnnounce;
 import com.streamr.client.protocol.message_layer.GroupKeyErrorResponse;
 import com.streamr.client.protocol.message_layer.GroupKeyRequest;
 import com.streamr.client.protocol.message_layer.GroupKeyResponse;
 import com.streamr.client.protocol.message_layer.Json;
+import com.streamr.client.protocol.message_layer.MalformedMessageException;
 import com.streamr.client.protocol.message_layer.MessageID;
 import com.streamr.client.protocol.message_layer.MessageRef;
 import com.streamr.client.protocol.message_layer.StreamMessage;
